# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TrainFlow API — fitness app for personal trainers and students
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/trainer/login:
    post:
      operationId: trainerLogin
      tags: [auth]
      summary: Trainer login with email + password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrainerLoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "401":
          description: Invalid credentials
  /auth/trainer/register:
    post:
      operationId: trainerRegister
      tags: [auth]
      summary: Register new trainer account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrainerRegisterInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
  /auth/student/magic:
    post:
      operationId: studentMagicLogin
      tags: [auth]
      summary: Student validates magic link token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MagicTokenInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "401":
          description: Invalid or expired token
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
        "401":
          description: Not authenticated
...
  /exercises:
    get:
      operationId: listExercises
      tags: [exercises]
      summary: List exercises (global + trainer's own)
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: muscleGroup
          in: query
          schema:
            type: string
        - name: scope
          in: query
          schema:
            type: string
            enum: [all, global, mine]
      responses:
        "200":
          description: Exercise list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Exercise"
    post:
      operationId: createExercise
      tags: [exercises]
      summary: Create custom exercise
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExerciseInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exercise"
  /exercises/{exerciseId}:
    get:
      operationId: getExercise
      tags: [exercises]
      summary: Get exercise detail
      parameters:
        - name: exerciseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Exercise
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exercise"

  # ─── TRAINING PLANS ──────────────────────────────────────────────────────────
...
  /conversations:
    get:
      operationId: listConversations
      tags: [chat]
      summary: List all conversations for current user
      responses:
        "200":
          description: Conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Conversation"
    post:
      operationId: createConversation
      tags: [chat]
      summary: Create direct conversation (trainer with student)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConversationInput"
      responses:
        "201":
          description: Conversation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
  /conversations/{conversationId}/messages:
    get:
      operationId: listMessages
      tags: [chat]
      summary: List messages in a conversation
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      tags: [chat]
      summary: Send a message
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"

# ─── COMPONENTS ────────────────────────────────────────────────────────────────
components:
  schemas:
...
    TrainerLoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    TrainerRegisterInput:
      type: object
      required: [fullName, email, password]
      properties:
        fullName:
          type: string
        email:
          type: string
        password:
          type: string
    MagicTokenInput:
      type: object
      required: [token]
      properties:
        token:
          type: string
    AuthResult:
      type: object
      required: [user, role]
      properties:
        user:
          $ref: "#/components/schemas/Profile"
        role:
          type: string
          enum: [trainer, student]
        isFirstLogin:
          type: boolean
    Profile:
      type: object
      required: [id, role, fullName, email]
      properties:
        id:
          type: integer
        role:
          type: string
          enum: [trainer, student]
        fullName:
          type: string
        email:
          type: string
        avatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string

    # Trainer Dashboard
...
    ExerciseInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        muscleGroup:
          type: ["string", "null"]
        equipment:
          type: ["string", "null"]
        gifUrl:
          type: ["string", "null"]
    Exercise:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        muscleGroup:
          type: ["string", "null"]
        equipment:
          type: ["string", "null"]
        gifUrl:
          type: ["string", "null"]
        isGlobal:
          type: boolean
        trainerId:
          type: ["integer", "null"]

    # Training Plans
...
    ConversationInput:
      type: object
      required: [studentId]
      properties:
        studentId:
          type: integer
    Conversation:
      type: object
      required: [id, type, trainerId, createdAt, unreadCount]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [direct]
        name:
          type: ["string", "null"]
        trainerId:
          type: integer
        otherUserName:
          type: string
        otherUserAvatarUrl:
          type: ["string", "null"]
        lastMessage:
          type: ["string", "null"]
        lastMessageAt:
          type: ["string", "null"]
        unreadCount:
          type: integer
        createdAt:
          type: string
    MessageInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
    Message:
      type: object
      required: [id, conversationId, senderId, content, sentAt]
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        senderId:
          type: integer
        senderName:
          type: string
        content:
          type: string
        sentAt:
          type: string
        readAt:
          type: ["string", "null"]