# lib/api-zod/src/generated/types/authResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TrainFlow API — fitness app for personal trainers and students
 * OpenAPI spec version: 0.1.0
 */
import type { AuthResultRole } from './authResultRole';
import type { Profile } from './profile';

export interface AuthResult {
  user: Profile;
  role: AuthResultRole;
  isFirstLogin?: boolean;
}